requirements.txt
yfinance==0.2.55
pandas>=2.2
matplotlib==3.9.2
feedparser==6.0.11
requests==2.32.3
google-generativeai==0.8.3
